Prepare for dropping Travis Linux builds and last of doc fixes.
Because the compiler on Trusty doesn't accept the -std=c++14 flag we are
dropping Travis CI Linux builds. The build actually does work on their systems
just not without specially pasing an old flag that enables behavior that
technically not standards compliant. Until C++14 std::array::operator[] is
not declared constexpr and the Apply build caught this but the GCC build did
not.

The preparation also included added Jenkins build badges to the README.md