Updated Jenkinsfile to use only the new pipeline syntax.
As part of this I move the build and test steps together for Debug and Release. This will give us slightly better parallelization. Since blue ocean now works, this will also let us see on cmake run along with its build in a single view on the blue ocean page.

I also removed the build status badge from the readme, because it wasn't economical to get that working now.