Fix jenkins2018 (#32)
* I am attempting to simplify the emscripten build so it works on Jenkins

* Added sourcing the emsdk_env script to fix emsripten build

* Attempting to set script vars on Jenkins node.

* Removed Jenkins Email

* Experimenting with Junit test results

* Another experiment with XML test results

* More tests with aggregating results

* Removing busted mail line

* Another experiment with Jenkins to see what we can get it to do.

* Fix an apparent syntax error.

* Another trial at getting the news jenkins stuff to do interesting things
* An experiment with the Jenkinsfile

* Another trial at getting the news jenkins stuff to do interesting things

* Attempt to remove build badge from Readme

* Updated Jenkinsfile to use only the new pipeline syntax.

As part of this I move the build and test steps together for Debug and Release. This will give us slightly better parallelization. Since blue ocean now works, this will also let us see on cmake run along with its build in a single view on the blue ocean page.

I also removed the build status badge from the readme, because it wasn't economical to get that working now.

* Moved Environment variable configuration to Jenkins nodes

* Attempt to fix status badge.

* Fix alt text on status badge.