Fix emscripten build in Jenkins CI
We didn't actually specify that we wanted to use em++ as the C++ compiler or emcc as the c
compiler. So when the build dir was cleared and we reran the build this caused the output to
be created with a normal compiler. Then we took the normal binary and tried to force it into
a javascript interpretter. Cleary this could never work.